Oh, it's just like .dlls but they call it something different on Linux. Fair enough.
USB Control working.  Iso still WIP, but if it works first time it
should be up and running tomorrow!

+++ Desktop Interface/genericusbdriver.h
@@ -35,7 +35,7 @@ public:
     virtual char *isoRead(unsigned int *newLength) = 0;
     void setBufferPtr(bufferControl *newPtr);
     void saveState(int *_out_deviceMode, double *_out_scopeGain, double *_out_currentPsuVoltage, int *_out_digitalPinState);
-    virtual void usbSendControl(int RequestType, int Request, int Value, int Index, int Length, unsigned char *LDATA) = 0;
+    virtual void usbSendControl(uint8_t RequestType, uint8_t Request, uint16_t Value, uint16_t Index, uint16_t Length, unsigned char *LDATA) = 0;
 protected:
     //State Vars
     unsigned char fGenTriple=0;
